Work in progress code for converting multiple size ranges as well as the VAE to TensorRT

+++ sdkit/utils/convert_model_utils.py
@@ -7,6 +7,89 @@ from sdkit.utils import log
 
 
 def convert_pipeline_unet_to_onnx(pipeline, save_path, opset=17, device=None, fp16: bool = False):
+    if os.path.exists(save_path) and os.stat(save_path).st_size > 0:
+        return
+
+    import torch
+
+    log.info("Making intermediate Unet ONNX..")
+
+    unet_in_channels = pipeline.unet.config.in_channels
+    unet_sample_size = pipeline.unet.config.sample_size
+    num_tokens = pipeline.text_encoder.config.max_position_embeddings
+    text_hidden_size = pipeline.text_encoder.config.hidden_size
+
+    model_args = (
+        torch.randn(2, unet_in_channels, unet_sample_size, unet_sample_size),
+        torch.randn(2),
+        torch.randn(2, num_tokens, text_hidden_size),
+        False,
+    )
+    input_names = ["sample", "timestep", "encoder_hidden_states", "return_dict"]
+    dynamic_axes = {
+        "sample": {0: "batch", 1: "channels", 2: "height", 3: "width"},
+        "timestep": {0: "batch"},
+        "encoder_hidden_states": {0: "batch", 1: "sequence"},
+    }
+    _convert_pipeline_model_to_onnx(
+        pipeline,
+        pipeline.unet,
+        model_args,
+        input_names,
+        dynamic_axes,
+        use_external_data_format=True,
+        save_path=save_path,
+        opset=opset,
+        device=device,
+        fp16=fp16,
+    )
+
+
+def convert_pipeline_vae_to_onnx(pipeline, save_path, opset=17, device=None, fp16: bool = False):
+    if os.path.exists(save_path) and os.stat(save_path).st_size > 0:
+        return
+
+    import torch
+
+    log.info("Making intermediate VAE ONNX..")
+
+    unet_in_channels = pipeline.unet.config.in_channels
+    unet_sample_size = pipeline.unet.config.sample_size
+
+    model_args = (
+        torch.randn(2, unet_in_channels, unet_sample_size, unet_sample_size),
+        False,
+    )
+    input_names = ["sample"]
+    dynamic_axes = {
+        "sample": {0: "batch", 1: "channels", 2: "height", 3: "width"},
+    }
+    _convert_pipeline_model_to_onnx(
+        pipeline,
+        pipeline.vae.decoder,
+        model_args,
+        input_names,
+        dynamic_axes,
+        use_external_data_format=False,
+        save_path=save_path,
+        opset=opset,
+        device=device,
+        fp16=fp16,
+    )
+
+
+def _convert_pipeline_model_to_onnx(
+    pipeline,
+    model,
+    model_args,
+    input_names,
+    dynamic_axes,
+    use_external_data_format,
+    save_path,
+    opset=17,
+    device=None,
+    fp16: bool = False,
+):
     import torch
     import onnx
     from torch.jit import TracerWarning
@@ -22,58 +105,48 @@ def convert_pipeline_unet_to_onnx(pipeline, save_path, opset=17, device=None, fp
         message="The shape inference of prim::Constant type is missing",
     )
 
-    unet_in_channels = pipeline.unet.config.in_channels
-    unet_sample_size = pipeline.unet.config.sample_size
-    num_tokens = pipeline.text_encoder.config.max_position_embeddings
-    text_hidden_size = pipeline.text_encoder.config.hidden_size
-
     orig_device = pipeline.device
-    orig_dtype = pipeline.unet.dtype
+    orig_dtype = pipeline.vae.dtype
 
     _dtype = torch.float16 if fp16 else torch.float32
-    _device = device if device else pipe.device
+    _device = device if device else pipeline.device
     pipeline = pipeline.to(_device, torch_dtype=_dtype)
 
-    tmp_dir = save_path + "_"  # collect the individual weights here
-    if os.path.exists(tmp_dir):
-        shutil.rmtree(tmp_dir)
-    os.mkdir(tmp_dir)
-    tmp_model_path = os.path.join(tmp_dir, "model.onnx")
+    if use_external_data_format:
+        tmp_dir = save_path + "_"  # collect the individual weights here
+        if os.path.exists(tmp_dir):
+            shutil.rmtree(tmp_dir)
+        os.mkdir(tmp_dir)
+    model_path = os.path.join(tmp_dir, "model.onnx") if use_external_data_format else save_path
 
     model_name, _ = os.path.splitext(save_path)
     model_name = os.path.basename(model_name)
 
+    model_args = tuple(m.to(device=_device, dtype=_dtype) for m in model_args if isinstance(m, torch.Tensor))
+
     onnx_export(
-        pipeline.unet,
-        model_args=(
-            torch.randn(2, unet_in_channels, unet_sample_size, unet_sample_size).to(device=_device, dtype=_dtype),
-            torch.randn(2).to(device=_device, dtype=_dtype),
-            torch.randn(2, num_tokens, text_hidden_size).to(device=_device, dtype=_dtype),
-            False,
-        ),
-        output_path=tmp_model_path,
-        ordered_input_names=["sample", "timestep", "encoder_hidden_states", "return_dict"],
+        model,
+        model_args=model_args,
+        output_path=model_path,
+        ordered_input_names=input_names,
         output_names=["out_sample"],  # has to be different from "sample" for correct tracing
-        dynamic_axes={
-            "sample": {0: "batch", 1: "channels", 2: "height", 3: "width"},
-            "timestep": {0: "batch"},
-            "encoder_hidden_states": {0: "batch", 1: "sequence"},
-        },
+        dynamic_axes=dynamic_axes,
         opset=opset,
-        use_external_data_format=True,  # UNet is > 2GB, so the weights need to be split
+        use_external_data_format=use_external_data_format,
     )
 
-    unet = onnx.load(tmp_model_path)
-    shutil.rmtree(tmp_dir)
-    # collate external tensor files into one
-    onnx.save_model(
-        unet,
-        save_path,
-        save_as_external_data=True,
-        all_tensors_to_one_file=True,
-        location=model_name + ".onnx_weights.pb",
-        convert_attribute=False,
-    )
+    if use_external_data_format:
+        model = onnx.load(model_path)
+        shutil.rmtree(tmp_dir)
+        # collate external tensor files into one
+        onnx.save_model(
+            model,
+            save_path,
+            save_as_external_data=use_external_data_format,
+            all_tensors_to_one_file=True,
+            location=model_name + ".onnx_weights.pb",
+            convert_attribute=False,
+        )
 
     pipeline = pipeline.to(orig_device, torch_dtype=orig_dtype)
 
@@ -110,17 +183,59 @@ def onnx_export(
     torch.onnx.export(model, model_args, output_path, **kwargs)
 
 
-def convert_onnx_unet_to_tensorrt(pipeline, onnx_path, save_path):
+def convert_onnx_unet_to_tensorrt(pipeline, onnx_path, trt_out_dir):
+    batch_size = 1
+    unet_in_channels = pipeline.unet.config.in_channels
+    num_tokens = pipeline.text_encoder.config.max_position_embeddings
+    text_hidden_size = pipeline.text_encoder.config.hidden_size
+
+    def get_shapes(min_size, max_size):
+        min_shape = {
+            "sample": (batch_size, unet_in_channels, min_size // 8, min_size // 8),
+            "encoder_hidden_states": (batch_size, num_tokens, text_hidden_size),
+            "timestep": (batch_size,),
+        }
+        max_shape = {
+            "sample": (batch_size * 2, unet_in_channels, max_size // 8, max_size // 8),
+            "encoder_hidden_states": (batch_size * 2, num_tokens, text_hidden_size),
+            "timestep": (batch_size * 2,),
+        }
+        return min_shape, max_shape
+
+    _convert_onnx_to_tensorrt(onnx_path, trt_out_dir, get_shapes, "unet")
+
+
+def convert_onnx_vae_to_tensorrt(pipeline, onnx_path, trt_out_dir):
+    batch_size = 1
+    unet_in_channels = pipeline.unet.config.in_channels
+
+    def get_shapes(min_size, max_size):
+        min_shape = {
+            "sample": (batch_size, unet_in_channels, min_size // 8, min_size // 8),
+        }
+        max_shape = {
+            "sample": (batch_size * 2, unet_in_channels, max_size // 8, max_size // 8),
+        }
+        return min_shape, max_shape
+
+    _convert_onnx_to_tensorrt(onnx_path, trt_out_dir, get_shapes, "vae")
+
+
+def _convert_onnx_to_tensorrt(onnx_path, trt_out_dir, shape_fn, name):
+    convert = False
+    for min_size, max_size in [(512, 768), (768, 1024), (1024, 1280)]:
+        save_path = os.path.join(trt_out_dir, f"{min_size}_{max_size}.trt")
+        if not os.path.exists(save_path) or os.stat(save_path).st_size == 0:
+            convert = True
+            break
+
+    if not convert:
+        return
+
     import tensorrt as trt
 
     TRT_LOGGER = trt.Logger(trt.Logger.INFO)
     TIMING_CACHE = "trt_timing.cache"
-
-    batch_size = 1
-    unet_in_channels = pipeline.unet.config.in_channels
-    unet_sample_size = pipeline.unet.config.sample_size
-    num_tokens = pipeline.text_encoder.config.max_position_embeddings
-    text_hidden_size = pipeline.text_encoder.config.hidden_size
 
     TRT_BUILDER = trt.Builder(TRT_LOGGER)
     network = TRT_BUILDER.create_network(1 << int(trt.NetworkDefinitionCreationFlag.EXPLICIT_BATCH))
@@ -132,60 +247,71 @@ def convert_onnx_unet_to_tensorrt(pipeline, onnx_path, save_path):
     if not parse_success:
         raise RuntimeError("ONNX model parsing failed")
 
-    config = TRT_BUILDER.create_builder_config()
-    profile = TRT_BUILDER.create_optimization_profile()
+    for min_size, max_size in [(512, 768), (768, 1024), (1024, 1280)]:
+        save_path = os.path.join(trt_out_dir, f"{min_size}_{max_size}.trt")
+        if os.path.exists(save_path) and os.stat(save_path).st_size > 0:
+            continue
 
-    if os.path.exists(TIMING_CACHE):
-        with open(TIMING_CACHE, "rb") as f:
-            timing_cache = config.create_timing_cache(f.read())
-    else:
-        timing_cache = config.create_timing_cache(b"")
-    config.set_timing_cache(timing_cache, ignore_mismatch=True)
+        log.info(f"Making TRT engine for {name}, size range from {min_size}x{min_size} to {max_size}x{max_size}..")
+        config = TRT_BUILDER.create_builder_config()
+        profile = TRT_BUILDER.create_optimization_profile()
 
-    min_shape = {
-        "sample": (batch_size, unet_in_channels, unet_sample_size, unet_sample_size),
-        "encoder_hidden_states": (batch_size, num_tokens, text_hidden_size),
-        "timestep": (batch_size,),
-    }
-    max_shape = {
-        "sample": (batch_size * 2, unet_in_channels, unet_sample_size, unet_sample_size),
-        "encoder_hidden_states": (batch_size * 2, num_tokens, text_hidden_size),
-        "timestep": (batch_size * 2,),
-    }
+        if os.path.exists(TIMING_CACHE):
+            with open(TIMING_CACHE, "rb") as f:
+                timing_cache = config.create_timing_cache(f.read())
+        else:
+            timing_cache = config.create_timing_cache(b"")
+        config.set_timing_cache(timing_cache, ignore_mismatch=True)
 
-    for name in min_shape.keys():
-        profile.set_shape(name, min_shape[name], min_shape[name], max_shape[name])
+        min_shape, max_shape = shape_fn(min_size, max_size)
 
-    config.add_optimization_profile(profile)
+        for name in min_shape.keys():
+            profile.set_shape(name, min_shape[name], min_shape[name], max_shape[name])
 
-    # config.max_workspace_size = 4096 * (1 << 20)
-    config.set_flag(trt.BuilderFlag.FP16)
-    serialized_engine = TRT_BUILDER.build_serialized_network(network, config)
+        config.add_optimization_profile(profile)
 
-    ## save TRT engine
-    with open(save_path, "wb") as f:
-        f.write(serialized_engine)
+        # config.max_workspace_size = 4096 * (1 << 20)
+        config.set_flag(trt.BuilderFlag.FP16)
+        serialized_engine = TRT_BUILDER.build_serialized_network(network, config)
 
-    # save the timing cache
-    timing_cache = config.get_timing_cache()
-    with timing_cache.serialize() as buffer:
-        with open(TIMING_CACHE, "wb") as f:
-            f.write(buffer)
-            f.flush()
-            os.fsync(f)
-            log.info(f"Wrote TRT timing cache to {TIMING_CACHE}")
+        ## save TRT engine
+        if not os.path.exists(trt_out_dir):
+            os.makedirs(trt_out_dir, exist_ok=True)
 
-    log.info(f"TRT Engine saved to {save_path}")
+        with open(save_path, "wb") as f:
+            f.write(serialized_engine)
+
+        # save the timing cache
+        timing_cache = config.get_timing_cache()
+        with timing_cache.serialize() as buffer:
+            with open(TIMING_CACHE, "wb") as f:
+                f.write(buffer)
+                f.flush()
+                os.fsync(f)
+                log.info(f"Wrote TRT timing cache to {TIMING_CACHE}")
+
+        log.info(f"TRT Engine saved to {save_path}")
 
 
-def convert_pipeline_unet_to_tensorrt(pipeline, save_path, opset=17, fp16: bool = False):
-    onnx_path = save_path + ".onnx"
+def convert_pipeline_to_onnx(pipeline, save_path, opset=17, device=None, fp16: bool = False):
+    unet_onnx = os.path.join(save_path, "unet", "model.onnx")
+    vae_onnx = os.path.join(save_path, "vae", "model.onnx")
 
-    if not os.path.exists(save_path) or os.stat(save_path).st_size == 0:
-        if not os.path.exists(onnx_path) or os.stat(onnx_path).st_size == 0:
-            log.info("Making intermediate ONNX..")
-            convert_pipeline_unet_to_onnx(pipeline, onnx_path, opset, device="cpu", fp16=False)
+    convert_pipeline_unet_to_onnx(pipeline, unet_onnx, opset, device=device, fp16=fp16)
+    convert_pipeline_vae_to_onnx(pipeline, vae_onnx, opset, device=device, fp16=fp16)
 
-        log.info("Converting intermediate ONNX to TensorRT..")
-        convert_onnx_unet_to_tensorrt(pipeline, onnx_path, save_path)
-        os.remove(onnx_path)
+
+def convert_pipeline_to_tensorrt(pipeline, trt_dir_path, opset=17, fp16: bool = False):
+    unet_path = os.path.join(trt_dir_path, "unet")
+    vae_path = os.path.join(trt_dir_path, "vae")
+
+    os.makedirs(unet_path, exist_ok=True)
+    os.makedirs(vae_path, exist_ok=True)
+
+    convert_pipeline_to_onnx(pipeline, trt_dir_path, opset, device="cpu", fp16=False)
+
+    unet_onnx = os.path.join(trt_dir_path, "unet", "model.onnx")
+    vae_onnx = os.path.join(trt_dir_path, "vae", "model.onnx")
+
+    convert_onnx_unet_to_tensorrt(pipeline, unet_onnx, unet_path)
+    convert_onnx_vae_to_tensorrt(pipeline, vae_onnx, vae_path)
